refactor: 의존관계 구분 위한 폴더별 index 파일 추가
