+++ src/domain/constants/index.js
@@ -1,0 +1,1 @@
+export {PRIZE_MAP} from './prizeMap';
