+++ src/domain/getTotalBenefit.js
@@ -1,4 +1,4 @@
-import {PRIZE_MAP} from './constants/prizeMap';
+import {PRIZE_MAP} from './constants';
 
 export const getTotalBenefit = lottoResult => {
   return Object.entries(lottoResult).reduce((acc, [prize, count]) => {
