+++ src/utils/index.js
@@ -1,0 +1,4 @@
+export {getDividedInteger} from './getDividedInteger';
+export {getPercentage} from './getPercentage';
+export {splitNumbers} from './splitNumbers';
+export {validateOnlyNumber, VALIDATE_ONLY_NUMBER_ERROR} from './validateOnlyNumber';
