+++ src/view/index.js
@@ -1,0 +1,2 @@
+export {getLineInput} from './getLineInput';
+export {queryValidInput} from './queryValidInput';
